fix(front): rune type consistency and bytes literal parsing
- Change Rune literal lowering from i64 to i32 to match core::I32 type
- Add Rune case to annotation_to_type for proper type unification
- Fix parse_bytes_literal to handle raw byte strings with multiple hashes

+++ crates/tribute-front/src/ast_to_ir/lower.rs
@@ -153,8 +153,8 @@ fn lower_expr<'db>(
         }
 
         ExprKind::RuneLit(c) => {
-            // Rune is lowered as i32 (Unicode code point)
-            let op = block.op(arith::Const::i64(ctx.db, location, c as i64));
+            // Rune is lowered as i32 (Unicode code point, matching core::I32)
+            let op = block.op(arith::Const::i32(ctx.db, location, c as i32));
             Some(op.result(ctx.db))
         }
